parsing all usb-id types with tests
Adds generic `CgType` for id, name and `CgParentType` for ids with
children in build.rs. A corresponding `UsbId` and `UsbIdWithChildren` in
lib.rs allows a ToToken impl for these commonly used items. type alises
are used to differentiate parsing and usage.

To implement the `FromId` trait for each USB id in lib.rs, the `UsbId`
needs to be made unique for each alias. I first attempted this with:

```
pub struct UsbId<Tag, T> {
      id: T,
          name: &'static str,
              _tag: std::marker::PhantomData<Tag>
}
```

I failed to figure out a way to add the add in the `ToToken` however.
Instead an associated constant is used. It serves the same purpose but
is not quite as clear, since the u8 values must be unique rather than a
`struct AtTag` for example.