optimise header check for min unique chars
Matching the whole comment line for next state isn't pretty but seems to
be the most robust way to detect this. See commented code removed based
on expected file ordering - this would be a prefered option but '#'
appears outside of section headers so it does not work.

This optimises the matching slightly by only checking the mininum num of
chars for the header description. I wanted to just use the '# TAG' but
'# HID Descriptor...' is a comment within the '# HCC' section.

Could think of other ways but I think this is simple and good enough.