Adds all USB IDs in the repository (#50)
* parsing all usb-id types with tests

Adds generic `CgType` for id, name and `CgParentType` for ids with
children in build.rs. A corresponding `UsbId` and `UsbIdWithChildren` in
lib.rs allows a ToToken impl for these commonly used items. type alises
are used to differentiate parsing and usage.

To implement the `FromId` trait for each USB id in lib.rs, the `UsbId`
needs to be made unique for each alias. I first attempted this with:

```
pub struct UsbId<Tag, T> {
      id: T,
          name: &'static str,
              _tag: std::marker::PhantomData<Tag>
}
```

I failed to figure out a way to add the add in the `ToToken` however.
Instead an associated constant is used. It serves the same purpose but
is not quite as clear, since the u8 values must be unique rather than a
`struct AtTag` for example.

* comments, fmt and tidy

* optimise header check for min unique chars

Matching the whole comment line for next state isn't pretty but seems to
be the most robust way to detect this. See commented code removed based
on expected file ordering - this would be a prefered option but '#'
appears outside of section headers so it does not work.

This optimises the matching slightly by only checking the mininum num of
chars for the header description. I wanted to just use the '# TAG' but
'# HID Descriptor...' is a comment within the '# HCC' section.

Could think of other ways but I think this is simple and good enough.

* make state process concise by using expect